fix(java): Move decompression code to it's own code for java
